i#7230 simrefs: Change skip_refs to not count markers
Changes the -skip_refs drmemtrace option to not count markers, matching the
-sim_refs and -warmup_refs options.  (The new -skip_records counts markers.)

Adds a changelist entry.

Updates the docs for -skip_refs, -sim_refs, and -warmup_refs to
clarify that they do not count markers.

Augments the drcachesim unit tests to check markers.

Cleanups up the cache and tlb simulators to not short-circuit on a
marker and skip later marker code, which requires a different solution
for core_for_thread handling of early markers and setting the tid in
tests.

A separate change will remove -skip_refs and -sim_refs from the view tool.

Issue: #7230