i#7230 simrefs: Add -exit_after_records drmemtrace option (#7555)
Adds a new drmemtrace option -exit_after_records which applies to all
tools and exits after N records. However, it does not work with dr$sim's
-warmup_fraction, and so we do not alias -sim_refs to this new option
but keep -sim_refs separate and counted inside the simulators (and view
tool: we could remove it there but without aliasing the old option it is
simpler to leave it).

Updates the -sim_refs option docs to reflect #7549's exit and to refer
to -exit_after_records for a similar option that works with all tools.

Adds get_output_record_ordinal() to the drmemtrace scheduler, required
to get the record count *after* -skip_instrs.

Adds a test that verifies multiple tools all exit and that this starts
counting after -skip_instrs.

Adds an entry to the release notes.

Issue: #7230