i#7230 simrefs: Add -skip_records drmemtrace option
Adds a new -skip_records drmemtrace option, implemented in the
framework for all tools.  This is a linear skip for simplicity: it
does not try to jump chunks for zipfiles.  It also does not re-display
the last timestamp,cpuid pair, again for simplicity.

Updates the view tool:
+ Do not print version + filetype if we skipped ahead.
+ Print a delayed timestamp on exit for -exit_after_records.

Adds an entry to the changelist.

Adds a test.

Issue: #7230