AArch64: Set unused parts of sve_context to zero. (#7558)
It was observed that all the signal tests were failing with non-debug
builds on SVE hardware with recent kernels: rt_sigreturn was causing a
segfault (which is confusing to debug because the PC value is taken from
the sigframe). The kernel was generating the segfault because struct
sve_context now has a "flags" member immediately after the "vl" member
and this field contained nonsense. Rather than add the "flags" member we
set the entire struct to zero before setting the fields we know about,
which is what we should have done in the first place: a field called
"__reserved" should generally be set to zero rather than left
uninitialised. We will probably add the "flags" member later.