i#7562: Properly handle redirected libc open() (#7563)
Fixes the private loader's redirect of libc open() to a new function
that properly handles the flags, instead of the old redirect which
treated them as DR's custom flags.

Adds a test that hits the truncation described in the issue and so fails
without the fix.

Also tested on the /dev/shm scenario described on the users list
https://groups.google.com/g/dynamorio-users/c/kZHTGiIUn8Y/m/rSurDOWeAQAJ.

Fixes #7562