i#7508: Call dr_redirect_execution on Linux after calling dr_nudge_client. (#7571)
#7509 introduced a new flag -memdump_on_window to capture memory dump
when a trace window opens. There is a significant delay after a nudge is
sent to the time it is received by the client.

Based on the dr_nudge_client comment:
```
On Linux, the nudge will not be delivered until this thread exits
the code cache.  Thus, if this routine is called from a clean call,
dr_redirect_execution() should be used to ensure cache exit.
```

hit_instr_count_threshold() itself is a clean call. And it is also
called by another clean call check_instr_count_threshold(). In both
cases, dr_redirect_execution() should be called at the end.

Manual test:
- without the change, memory dump files are created minutes after a
window is opened (based on the corresponding trace file timstamp).
- with the change, the timestamps of the memory dump file and the trace
file in the same window is close.

Issue: #7508