i#7576 Use correct context when detaching
Nudge-triggered detach on Linux appears to be broken because it is
switching to the wrong context in the final step.

Previously the nudge handler would save the signal context from the
nudge signal and switch to that when going native, but by the time the
nudge is handled the signal has returned and `nudged_sigctx` is a
dangling pointer.

Even if we fix the dangling pointer by making a copy of the signal
context on the heap, the signal context is still the wrong context to
use. If the signal context PC is an address in the code cache DynamoRIO
should be able to translate the state and switch to it when detaching,
but it might not be. In my testing the signal context PC was always in
libdynamorio.so code. `detach_externally_on_new_stack()` would fail to
translate the context and try to jump into libdynamorio.so code and
crash.

Instead `detach_externally_on_new_stack()` should just use the mcontext
for the thread. With this change DynamoRIO appears to be able to detach
from the process now, but not reliably. Sometimes it succeeds, sometimes
it crashes in DynamoRIO code during detach, sometimes it detaches then
crashes in the native app code. There are obviously more bugs to find in
this area but this is progress.

Issue: #7576