i#7230 sim refs: Remove -skip_refs and -sim_refs from view tool
Removes support for -skip_refs and -sim_refs from the view tool.
These now count only non-markers and are limited to the simulator tools.
They have been replaced by analyzer-provided -skip_records and
-exit_after_records.

Adds an error message if a user tries to use the old options.

Removes the delay for printing version and filetype records. It is
from a very old file reader which did not re-order records; the modern
file reader re-orders so the reader_t always sees the tid before the
first visible record.  This avoids having to port the old sim_refs
logic for the delayed headers.

Augments invariant_checker with a flag "trace_incomplete" which
relaxes the invariant that a thread exit is seen and also adds
handling of skip_records via a heuristic.

Adjusts the drcacheoff.skip test to expect 2 more instructions, as
-exit_after_records does not count synthetic records including the 2
markers inserted after -skip_instrs.

Augments the view_test's testing of skipping and limiting to create an
analyzer and use the new options.

Changes the one_early test to use the TLB simulator instead of the
view tool.

Fixes #7230