Replace FIXME with XXX and TODO
Updates all the "FIXME" comments in the repository to instead use
"TODO" (for those about missing implementations: "NYI" or
"ASSERT_NOT_IMPLEMENTED") or "XXX" (everything else).  This matches
our style guide which discourages "FIXME" as its connotations are too
negative and too easily mis-interpreted by readers and in code audits
to imply that the current state of things is broken in some way
instead of just meaning something could be improved.