i#7574 abandoned br tar: Show scheduling behavior (#7579)
Adds a new scheduler unit test that demonstrates how an abandoned
TRACE_MARKER_TYPE_BRANCH_TARGET is treated by the drmemtrace scheduler
logic.

When there's an abandoned branch_target marker in the previous chunk,
and other unrelated markers intervening before the corresponding
encoding+instr in the next chunk, it affects the scheduler in unexpected
ways: the instr ordinals are erroneously incremented for just the
branch_target and then again at the encoding+instr, and there may be a
context switch that splits up the branch_target marker and its
corresponding instruction.

The plan is to ensure branch_target markers are not abandoned in
raw2trace, and to also add a scheduler workaround to properly handle
older traces that do have this problem.

Issue: #7574