i#7504 time scale: Add -scale_timeouts drmemtrace option (#7583)
Adds a drmemtrace option -scale_timeouts which leverages the new drx
sleep scaling feature.

Adds a test which compares the sleep count in a drmemtrace offline trace
with and without scaling. It was difficult to make the test robust
across different loaded test machines so it uses an adaptive duration
approach.

The test occasionally hits #7582 which is fixed here (not separated out
as the convenient test is here) by dropping the 3rd pending instance of
any non-real-time signal; tested by running the test 100x and seeing no
failures (previously it would fail every 3 to 10 runs or so):

```
$ ctest --repeat-until-fail 100 -R burst_sleep
...
100% tests passed, 0 tests failed out of 1
Total Test time (real) = 281.66 sec
```

Issue: #7504, #7582
Fixes #7582