i#7574 abandoned br tar: Fix issue in raw2trace (#7584)
Fixes an issue in raw2trace's handling of chunks that caused the
TRACE_MARKER_TYPE_BRANCH_TARGET for certain indirect branches to be left
behind in the prior chunk.

This happened when all of the following conditions were true:
- the indirect branch was the first instruction of a new chunk
- the indirect branch was a part of a sequence of delayed branches that
crossed over chunk boundary
- the indirect branch had a prior instance in the previous chunk so that
an encoding entry was not initially added before it (but would be added
by raw2trace upon opening the new chunk)

Adds a unit test that reproduces the issue, which fails without this
fix.

Issue: #7574