i#5771: On AArch64 optionally mangle MRS to make CTR_EL0.DIC seem zero. (#7578)
This is a work-around rather than a proper fix for issue #5771. It may
also facilitate experimentation by users.

When "-fake_ctr_dic" is specified the MRS instruction is mangled by
adding a following AND (immediate) instruction that clears the
corresponding bit. Most programs will then execute the IC IVAU
instruction that DynamoRIO currently relies on for detecting a code
modification. (See, for example, LLVM's implementation of
__clear_cache.)

Issue: #5771