i#7588 drsyscall: Add timestamp to syscall beginning and end records. (#7586)
Add new DrSyscall record types DRSYS_SYSCALL_NUMBER_TIMESTAMP and
DRSYS_RECORD_END_TIMESTAMP to replace DRSYS_SYSCALL_NUMBER and
DRSYS_RECORD_END. The new record types have a timestamp to record when a
syscall is called and ended.

Add START_PACKED_STRUCTURE and END_PACKED_STRUCTURE to the union in
syscall_record_t which is needed
to ensure the struct is packed.

The change increase the size of syscall_record_t for 32-bit, so it's not
backward compatible for 32-bit. The size of syscall_record_t for 64-bit
stays the same.

Update syscall-records-test to look for timestamps.

The timestamp enables us to compute how long does a syscall take, and
filter syscall records based on timestamps.

Issue: #7588 .