i#7504 time scale: Increase zero sleep to scale it
A sleep system call with a zero duration is not a nop and does block.
We increase a requested zero to 10us so it will be inflated too.

Adds a sanity check on the nop stats to the drx sleep test.
Manually tested further on a large application with drmemtrace
-scale_timeouts.

Issue: #7504