i#7504 time scale: Fix EINTR test (#7593)
Fixes a problem in the drx sleep scaling test when checking the
remaining time on EINTR where it was only checking the seconds field. We
have to allow an larger remainder than the original, observed even
without scaling and due to the kernel rounding up to underlying clock
granularities as noted in the Linux man pages.