i#7576 nudge detach: Use mcontext instead of signal context (#7581)
Nudge-triggered detach on Linux appears to be broken because it is
switching to the wrong context in the final step.

The initial symptom I saw was an assert firing during the detach
process:
```core/unix/signal.c:2826 mc != NULL && sc != NULL```

The nudge handler would save the signal context from the nudge signal and switch to that when going native, but by the time the nudge is handled the signal has returned and `nudged_sigcxt` is a dangling pointer. The 

Even if we fix the dangling pointer by making a copy of the signal context on the heap, the signal context is still the wrong context to use. If the signal context PC is an address in the code cache DynamoRIO should be able to translate the state and switch to it when detaching, but it might not be. In my testing the signal context PC was always in libdynamorio.so code. `detach_externally_on_new_stack()` would fail to translate the context and try to jump into libdynamorio.so code and crash.

Instead `detach_externally_on_new_stack()` should just use the mcontext for the thread. With that change it appears to be detaching reliably. I have tested it with some simple apps like `linux.infloop` and some more complicated apps like `python3`, with and without logging, and with and without drmemtrace.

The existing `client.detach_test` regression test don't seem to reproduce this issue unless you use a debug build of DynamoRIO with logging turned on (`-loglevel 2` or greater). There is an internal detach test `api.detach_state` which systematically tests app state is consistent. We need to adapt this to test external detach as well but this will come in later commits.

Issue: #7576