i#7504 scale time: Fix missing sleep filter (#7599)
Adds sleep syscalls to the filter event which were omitted by PR #7568.
This did not affect the drx_sleep_scale-test as its use of syscall()
caused DR to not find the number, and it seems that the libc nanosleep()
wrapper ends up also using a split-block syscall, but it could affect
some uses. (The filter only matters when DR can statically determine the
syscall number, in which case if no client or DR itself wants to observe
that syscall then DR will inline it into the cache with no pre or post
actions.)

Adds an inlined-syscall test (x86-64 only for simplicity) which fails
w/o the filter.

Issue: #7504