i#7604 user data: Add drmgr syscall filter event
Adds a drmgr filter syscall event which replaces the base DR event and
provides priorities and user_data parameters.

Removes the lazy drmgr pre and post syscall registration as it is no
longer needed with drmgmr always registering a filter event, but
leaves the underlying lazy mechanism in place in case we want it for
something else.

Updates all our samples and tests to use the drmgr event.

Adds a note to the minor compatibility change list as this will break
user's builds, but has no binary compatibility consequence.

Issue: #7604