i#7496 eret continuity: Add kernel IBT read-ahead queue in next_record
Adds logic in the drmemtrace scheduler to set the branch target of syscall- and switch-final indirect branch instructions to the actual next instruction in the trace. This includes the following notable cases:
- when there is a syscall trace immediately followed by a switch trace. The syscall-final IBT must be set to the first instr in the switch sequence
- when the next input is not yet known after a blocking syscall. The syscall sequence includes all the idles, and the syscall-final IBT is returned only when the next input is figured out.

Replaces the indirect branch target (whether in the branch_target marker trace_entry_t, or an indirect branch instr memref_t) with a sentinel during injection. When this sentinel is the next one to be returned in next_record, we read-ahead records for that output until we see the target instruction. If the output stream returns any non-OK status during this readahead, we return that status and wait to return the queued-up records. If no instr is left in the output stream, we remove the indirect branch instr.

Adds a new get_input_instr_ordinal API, similar to the existing get_input_record_ordinal, which returns the input's adjusted ordinals for instructions. This is useful in the analyzer USE_INPUT_ORDINALS mode.

Modifies the existing record_type_is_indirect_branch_instr to separate out the setter functionality into a separate record_type_set_indirect_branch_instr.

Adds unit tests for:
- analyzer parallel mode with syscall injection
- scheduler syscall+switch injection that has some idles
- modifies existing syscall and switch scheduler tests

Issue: #7496