i#7576 detach: Check AArch64 SIMD registers in detach_state test (#7600)
Adds support for checking the AArch64 vector registers in the
api.detach_state test. When built with SVE support it will check all Z
and P registers and the FFR. When built without SVE support it will
check all the Neon V registers.

The SVE support is intended to work with any vector length and has been
tested with 128-bit and 256-bit vector lengths.

Issue: #7576, #4698