i#7598: Pause all threads at takeover
Adds a new DR option -synchronous_attach which is on by default.  When
on, during DR start when DR takes over all other threads is no longer
lets each one execute immediately.  Instead, each thread is suspended
at a barrier, and only once all threads are taken over and the
client's post-attach event is called are the threads allowed to
continue.

This option eliminates the gap between client snapshots and thread
execution, which was causing problems with missing data in clients.
Client snapshots need to be placed in the post-attach event to resolve
this issue.

The downside is that if there are many threads it can take a while to
take them all over via signal, so some may be paused at the barrier
for a while.  The runtime option can be disabled if this is a problem.

Adds documentation on the new behavior.

Updates drx to perform its timer snapshots in the post-attach event,
to avoid races and gaps.

Adds a test by splitting the drcacheoff.burst_threads_counts
-align_endpoints drmemtrace test into two: one with
-synchronous_attach where all 49 threads are always instrumented, and
one with the old scheme where most threads start early (staggered
attach) and are discarded by timestamp.

Future work may also try to tackle #1305 by moving the start point
with its thread takeover inside setup, and disallow a separate setup
from start.  For now, support for re-calling dr_app_start() before a
stop is kept, with the dr_attach_finished event reset which looks like
a possible error from before.

Issue: #7598, #1305