i#7496 eret continuity: Add get_next_trace_pc stream API with reader readahead
Adds a new memtrace_stream_t::get_next_trace_pc API that returns the next continuous pc that will be seen in the trace, either in an instr_t or in a TRACE_MARKER_TYPE_KERNEL_EVENT.

Adds functionality in reader_t and record_reader_t to read-ahead the trace_entry_t from the input so that we always know the next continuous pc in the trace.

The implementation is added in a trace_entry_readahead_helper_t class to allow it to be shared between reader_t and record_reader_t, which have similar but not the same interfaces. Also added entry_queue_t that holds the logic to dequeue up trace_entry_t records, which supports storing readahead and non-readahead entries, and also logic to tell whether there are enough readahead entries for us to know the next continuous pc in the trace.

Adds analyzer unit tests that verify the correct operation of get_next_trace_pc for memref_t and trace_entry_t tools. Also ensured that the trace skip related tests pass.

Issue: #7496