i#7552: Fix a race condition in attach-memory-dump-syscall-test. (#7623)
There is a race condition between the post-command to read the syscall
record file and the DR client in attach-memory-dump-syscall-test.dll.c
writing to the file. The post-command is triggered by writing "done" to
STDOUT or STDERR. This race is resolved by having the DR client write
the done only after it has finished writing to the syscall record file
within its exit event handler.

Removing print("done") from the signal_handler in
syscall-records-attach.c introduces another race condition.
A race exists between the application making syscalls and a SIGTERM that
is sent after the attach event. This could potentially cause the test to
fail if no syscalls are recorded before the signal is sent. The solution
is to explicitly perform a write syscall in the code, guaranteeing that
at least one syscall is recorded for the test.

Tested:
Without the changes, "ctest --repeat-until-fail 100 -R
attach-memory-dump-syscall-test" usually fails before hitting 100 tests.

With the changes, multiple runs of "ctest --repeat-until-fail 1000 -R
attach-memory-dump-syscall-test " have 0 failures.

Fixes #7552