i#7572: Fix mid-block-detach drmemtrace bug
Adds a new test drcacheoff.longblock which reproduces a bug where a
detach mid-block results in an incorrectly-complete post-processed
block missing load/store records in the tail of the block.  The test
fails with an invariant error on missing read records in HEAD.

Fixes the test by adding a new trace record
TRACE_MARKER_TYPE_MIDBLOCK_END_PC just before the footer in each
thread. This holds the thread's final PC value.  This marker is used
by raw2trace to identify a block interruption point, just like a
kernel event marker would, and the same
TRACE_MARKER_TYPE_UNCOMPLETED_INSTRUCTION is emitted as well (it
doesn't really help analyzers/simulators for detach, but it will help
for DR mid-block relocation: #5790).  The marker is removed and not
visible in the final trace.

This same marker can be used to solve block truncation for DR
mid-block relocation (#5790) once DR provides a relocation event.

Updates DR detach to set each thread's mcontext for retrieval by
clients in the thread exit event.  Documents that the pc field is in
fact set for the thread exit event (and verified it is set for regular
exit: points to exit syscall post-pc so the uncompleted marker does
not get added there which seems good).

Fixes #7572