i#7496 eret continuity: Add reader_base_t (#7629)
Adds a new base class for reader_t and record_reader_t, called
reader_base_t, which holds some interface definitions and implementation
common to the two kinds of readers.

Adds get_next_entry() which is intended to be used by derived classes to
get the next entry, which is different from read_next_entry() which
actually reads the next entry from the underlying trace source and is
now made private. get_next_entry() may or may not invoke
read_next_entry() depending on whether there are queued entries already
available.

Removes read_queued_entry() from all derived class implementations, as
that is now simply handled by the reader_base_t::get_next_entry()
itself.

Renamed eof_ to at_eof_ in record_reader_t and its derived classes, as
that is now declared in reader_base_t, shared with reader_t. at_eof_ is
without a getter and setter on purpose, to allow derived classes to more
easily use it.

Renamed cur_entry_ to entry_copy_ in record_reader_t and its derived
classes, as that is now declared in reader_base_t shared with reader_t.
This can be used to hold the entry read by the derived classes of
record_reader_t.

Requires derived classes to provide whether or not they operate in the
online mode to reader_base_t's constructor, instead of asking them to
set online_ by themselves.

Enables doxygen for private class members, as we want to be able to link
to read_next_entry() which is part of the virtual interface of
reader_base_t, though private.

The responsibility to read queued entries before accessing the actual
stream of trace entries is taken back by the reader_base_t in this PR.
It was moved to the individual types of readers in #5900, perhaps in an
effort to simply the reader itself.

This is in preparation for #7621 which adds more entry queuing logic to
reader_base_t.

Issue: #7496, #5727