i#7496 eret continuity: Handle sched readahead in get_next_trace_pc (#7633)
Modifies the get_next_trace_pc() implementation for the scheduler to
return the next trace pc accounting for the records read-ahead by the
scheduler. We keep track of selected stream state, specifically the
get_next_trace_pc() output, when adding records to the scheduler
read-ahead queue, in cases where we expect the reader to not provide an
accurate state later.

Adds a test that uses the invariant checker in the serial mode which
invokes the various scheduler readahead logic. The invariant checks for
get_next_trace_pc() fail without this fix.

Issue: #7496