i#7496 eret continuity: Add get_next_trace_pc reader readahead API (#7621)
Adds a new memtrace_stream_t::get_next_trace_pc() API that returns the
next continuous pc that will be seen in the trace. The next continuous
trace pc is either the pc of the next instr_t or the address held by the
next TRACE_MARKER_TYPE_KERNEL_EVENT, whichever comes first.

Adds functionality in reader_base_t to read-ahead as many trace_entry_t
records from the input as required so that we always know the next
continuous pc in the trace. This value is supplied via the
get_next_trace_pc() API.

Also added entry_queue_t that holds the logic to (de)queue trace_entry_t
records, which is used to support all kinds of read-ahead by the readers
now, and also logic to tell whether there are enough read-ahead entries
for us to know the next continuous pc in the trace.

Adds analyzer unit tests that verify the correct operation of
get_next_trace_pc for memref_t and trace_entry_t tools.

Adds new invariant checks that verify correctness of the next trace pc
received from the stream API.

Ensured that existing skip tests pass, and so do the existing invariant
checker tests on real traces.

Adds more skip tests that verify that the same instr pc is seen at the
same instr ordinals, when different combinations of initial sim and skip
values are used, which is especially important to verify zipfile
skipping.

memtrace_stream_t::get_next_trace_pc() is not currently supported in the
online analysis mode, and for i-filtered traces.

This is in preparation for #7614 which would use the new API to set the
PC for indirect branch targets at injected kernel trace ends, to the
next pc in the trace.

Issue: #7496