i#7496 eret continuity: Use get_next_trace_pc for kernel trace final branch
Adds logic in the drmemtrace scheduler to set the branch target of syscall- and switch-final indirect branch instructions (e.g., eret on AArch64) to the next instruction in the thread's trace.

- The syscall-trace-final indirect branch now points to the next instruction as seen on the thread shard, which may be the syscall-fallthrough or the syscall instruction again for auto-restart syscalls.
- The context-switch-trace-final indirect branch points to the next instruction of the new input.
- When there is a syscall trace immediately followed by a switch trace, the syscall-final indirect branch still points to next instruction as seen on the thread shard.

When reading the kernel trace templates, we discover and record the first instruction of the template, which we present as the next trace pc just prior to the trace's injection.

Augments the get_next_trace_pc() implementation to return the next pc in the injected trace, when it is queried in the middle of dynamic injection. Enables invariant checks on the get_next_trace_pc() output during the dynamically injected part of the trace.

Adds unit tests for:
- analyzer parallel mode with syscall injection
- scheduler syscall+switch injection that has some idles
- extends existing syscall and switch scheduler tests

Issue: #7496