i#7649 AArch64 detach_state: Fix sigreturn segfault
Porting the detach state thread_check_gprs_from_DR() test to AArch64
(#7627) highlighted an uninitialized data bug affecting
thread_set_self_mcontext().

thread_set_self_mcontext() creates a kernel_sigcontext_t struct on the
stack and populates it with the data from the mcontext then executes a
sigreturn to switch to that state. On AArch64 kernel_sigcontext_t has a
4096 byte array called __reserved which is used as storage for simd
state and other state added by later architecture extensions.

thread_set_self_mcontext() was leaving this array uninitialized and
relying on mcontext_to_sigcontext_simd() to fill it in, however a bug in
the logic of that function meant that list of context records was not
properly terminated on non-SVE hardware. Without the null terminator
record the kernel would read the uninitialized data past the end of the
context record list and the sigreturn would fail, delivering a SIGSEGV.

When I looked at mcontext_to_sigcontext_simd() it looked to me like
there were more problems that just the missing terminator.

The way it was written, it would always create fpsimd_context and
sve_context records, overwriting any data already present in the
__reserved storage. That probably doesn't matter when the sigcontext is
being used for sigreturn, but it is also used to fill in translated app
state when delivering a signal to the app. In the latter case it would
probably be better to preserve whatever context records are present to
pass on to the app's signal handler.

With this in mind I have rewritten mcontext_to_sigcontext_simd() to walk
the list of context records and find existing fpsimd and sve records so
we can write the SIMD register state without clobbering all the other
records. If fpsimd and sve records are not found in the list new ones
are created at the end.

Callers to mcontext_to_sigcontext() will need to ensure that the
__reserved[] data is initialized, either by creating the context
records, or by initializing the memory to 0 which effectively creates a
null terminator records. I have looked through the existing call sites
and I believe we are covered.

I have also started putting together a test covering AArch64 sigcontext
handling but will come in a future PR because it depends on other work I
have been doing to port and refactor the detach_state test.

Issue: #7649, #4698