i#7496 eret continuity: Set branch target using get_next_trace_pc() (#7639)
Adds logic in the drmemtrace scheduler to set the branch target of
syscall- and switch-final indirect branch instructions (e.g., eret on
AArch64) to the next instruction in the thread's trace.

We use the memtrace_stream_t::get_next_trace_pc() API to discover the
next trace pc that will be seen on a thread trace.

- The syscall-trace-final indirect branch now points to the next
instruction as seen on the thread shard, which may be the
syscall-fallthrough, the syscall instruction again for interrupted
auto-restart syscalls, or the signal return PC for sigreturn.
- The context-switch-trace-final indirect branch points to the next
instruction of the new input.
- When there is a syscall trace immediately followed by a switch trace,
the syscall-final indirect branch still points to next instruction as
seen on the old thread shard, as the decision to switch hasn't been made
by the scheduler yet.

When reading the kernel trace templates, we discover and record the
first instruction PC of the template, which we present as the next trace
PC just prior to the trace's injection.

Augments the get_next_trace_pc() implementation to return the next pc in
the injected trace, when it is queried in the middle of dynamic
injection.

Enables invariant checks on the get_next_trace_pc() output during the
dynamically injected part of the trace. Enables PC continuity checks in
the invariant_checker that verify control flow continuity of the
branch_target marker value to the post-syscall trace pc.

Adds unit tests to verify get_next_trace_pc() output, and
kernel-trace-final indirect branch target values in various modes of
operation:
- analyzer parallel mode with syscall injection
- scheduler syscall+switch injection that has a syscall trace followed
immediately by a switch trace
- extends existing syscall and switch scheduler tests with more checks.

Also manually ran the invariant checker on two large internal app traces
(x86 and aarch64), and confirmed that no unexpected invariant errors
were reported.

Issue: #7496