i#7644: Allow specifying a default syscall trace in template file
Allows the user to specify a default system call trace in the syscall template file. This trace will be used to inject syscall instances for syscalls that otherwise do not have any template specified in the syscall template file. This is to help with cases where the user wants to generate a trace with all syscalls injected, and not just the ones that we have a trace for. This default template only needs to contain the indirect branch (sysret or eret).

Adjusts invariant checks that were found to fail after such injection for all syscalls was enabled: specifically to check for pc continuity with the prior syscall-final branch target marker, even when the regular pc continuity checks were disabled. This cases existed before this PR, but was not seen because we didn't try to inject that particular syscall before.

Also updates the mock syscall trace template file to include a default trace, and ensures the existing tests using this file pass.

Issue: #7644