i#7643: Simplify icache sync on AArch64: act on IC IVAU immediately.
Instead of accumulating IC IVAU addresses in a global struct and only
acting on them when an ISB is seen, act on each IC IVAU immediately
and in the same thread. This may be slower in some cases as there may
be multiple calls to flush_fragments_synch_unlink_priv instead of a
single call but it avoids an inter-thread issue.

Fixes #7643

Change-Id: I01fe09f04a2a6332c0f0c2724093a4bb2d283cbd