i#6684: Adapt invariant_checker to core sharded traces
Adapts the drmemtrace invariant_checker tool to work on core-sharded traces; this includes dynamically core-sharded and core-sharded-on-disk traces both.

At a high level, we reset the invariant checker context when we see a context switch, which includes switches to a different thread and the same thread running after some core_idles.

For core-sharded traces injected with kernel syscall and switch sequences, we also verify that the kernel trace end indirect branch provides continuity to the next trace.

Issue: #6684