i#297: Fix client thread synch issues and api.static_sideline test (#7668)
Gets the api.static_sideline test working again and removes its _FLAKY
status:

+ Moves the test's wait to the post-attach event b/c client threads wait
for dr_app_start.

+ Passes in the dcontext to thread exit in cleanup_and_terminate to work
around the #3535 TLS ignoring so client threads can exit on their own:
but for x86 only uses it for client threads, to preserve the #3535
detach race fix.

+ Augments the safe-spot condition for a client thread to be unsafe if
dcontext->is_exiting is set, to avoid killing a thread that has
incremented exiting_thread_count (is_exiting is set right before doing
that) and is waiting on the thread_initexit_lock to exit. This causes
the synchall loop to retry and drop the lock, letting the client threads
exit reliably. This solves the static-DR issue of not being able to
distinguish the client from DR.

The test is disabled for AArch64 due to #7670.

Tested:
```
$ ctest --repeat-until-fail 500 -R static_sideline
100% tests passed, 0 tests failed out of 1
```

Fixes #297