i#6684: Adapt invariant_checker to core sharded traces (#7665)
Adapts the drmemtrace invariant_checker tool to work on core-sharded
traces; this includes dynamically core-sharded and core-sharded-on-disk
traces both.

At a high level, we reset some of the invariant checker state when we
see a context switch, which includes switches to a different thread, or
the same thread after some core_idles.

For core-sharded traces with kernel syscall and switch sequences, we
also verify that the kernel trace end indirect branch has continuity to
the coming up trace, and that all instances of syscalls and switches are
injected. Currently, we use OFFLINE_FILE_TYPE_KERNEL_SYSCALLS to mean
that both syscall and switch sequences should be expected. Will rename
that in an upcoming PR after some internal discussions.

Updates mock_switch_sequences.x64.zip to adhere to the requirement of an
indirect branch at the end of each sequence.

Adds unit tests to:
- verify invariants for mock_switch_sequences.x64.zip
- verify invariants for a checked-in core-sharded-on-disk trace
- verify invariants for a dynamically core-sharded trace with kernel and
switch sequences
- verify specific invariants in invariant_checker_test unit tests

Manually tested on two large internal core-sharded-on-disk-with-kernel
trace that no invariant errors are reported. Also verified that an
invariant error was reported for a user-mode-only trace that was
affected by missing branch target due to #7574.

Issue: #6684